:recycle: Refactor fills DTOs into separate submodules
